Refactor String -> RusType enum (tag/vec/wasm issues)
In order to better support tagged types, arrays, and conversion between
the wasm boundary instead of a simple String type there is now a
composable RustType enum to allow for more information at every stage
and better code generation as a result.

This allows us to not require the tag for paremters while stilly
serialization it for tagged types, as well as not requiring separate
types for vecs of primitive types.