Change metadata structs to return errors instead of options
This allows better structural validation from javascript as we can embed
the entire parsing code into a try/catch rather than manually checking
if the return is undefined. If the user simply wanted to check if it was
that type then they can still use `kind()` on metadatums, and for the
map we added a `has(key)` to cover that use-case.