TxBuilder: max output size + multiple change address support
Users with many NFTs could run into problems creating certain txs that
returned many of them as a change output as it would surpass the maximum
output size limit in the ledger. The TxBuilder now takes this limit into
account and errors if it is passed, and during the change output
creation will attempt to spread out the NFTs over multiple outputs to
not surpass the ledger limit.

This is done as a greedy selection for now as the limit is currently 4kb
or so and each asset is at most 64 bytes each so this should suffice.